Add skeleton file preprocessor; add directive rule to the grammar
Add skeleton file preprocessor and the skeleton file itself. Preprocessor
takes the output from the parser generator and inserts it into the skeleton
file, replacing the directives in it.

Add the grammar directives.
Directive may be placed before the rule and must start with the at sign '@'.
Currently, there is only one directive "entry", which marks the grammar rule
as the starting rule. This is needed for the preprocessor to insert the call
to the starting function. Currently, there is no check for the wrong
directives.