Add position reset; remove redundant savings
Generator will insert _reset at the end of the first part of an alternative,
if this alternative is not the last. This is needed, because otherwise the
next alternative will fail trying to match at the wrong position.

Generator won't insert `pos = self._mark()` at the beginning of a rule
and reset at the end, if the last alternative empty. Empty last alternative
means that this parsing function is a desugared ZeroOrOne quantifier,
which is true in any case. So there is no need to save and restore parsing
position.