Add metaname and metarule; add variables to parser generator
Add metanames and metarules to the grammar. Metaname is an optional identifier,
followed by an equality sign '=', which precedes a particle. It should be
disallowed to use metanames for lookahead particles. Metarule consists of
a percent sign '%', followed by an identifier. Metarule follows a sequence.
Corresponding members are added: metaname to Part and metarule to Alt.

Add variables to parser generator. Also add a workpiece for semantic actions.
Actually I really don't know how to implement semantic actions to make this
approach not so complicated and cross-language. Maybe I'll just make parsers
generate a parsing tree...

Update todo.txt.