Add new semantic actions format.
Rewrite TreeModifier's logic. If the semantic rule returns False, it will
not be called on the next tree traversal iteration. If the semantic rule
returns None, it will not be called anymore since the last call. Add max
iteration count protection to prevent infinite loops. This is a crutch, but
I cannot give up with something gracier.

Add MetaRef, MetaRule nodes.

Add reindent function to the python generator. I think, this function should
be moved into the utility module. I suppose that this function will be used
by all backends to properly indent semantic rule's contents, but for Python
it is essential.