Make left recursion almost work
Add DetectLeftRecursion modifier, LR class to nodes and _LR class to
the generated parser.

In the current approach the set of involved rules is created for each
recursive alternative in the rule. Context manager _setup_lr loads the
set into the parser and unloads on exit. But this approach leads to an
infinite recursion, that is, it doesn't work.

Why? Because when one context manager exits and another fires up, it
replaces the current left recursion state. And ta-daa! New and fresh
eval_set allows parser to not fail and repeat this branch, and then
it replaces exhausted eval_set again, and again, and again...