Add FilePreprocessor and change the way how files are passed to the program
FilePreprocessor class inherits from the Preprocessor and incapsulates
logic that is specific to file handling. It takes not directory with files
and directory to output, but an iterable of tuples or a mapping, that
says, which exact input file should be translated into which output file.

Change generator.py and __main__.py modules correspondingly. Also change
make.sh script.