Rewrite node module.
Grammar tree uses Doubly Linked List structure, which allows to modify
the tree in-place.

Remove Part node, add NamedItem instead. Make quantifiers and lookaheads
nested. The structure of the Alt is the following:

```
Alt
  NamedItem
    Lookahead (And, Not)
      Quantifier (ZeroOrOne, ZeroOrMore, OneOrMore, Repetition)
        Item (String, Char, Id, etc.)
```

Rewrite TreeModifier. The concept of "stages" was better in thoughts then
in code, so I removed it.

It seems I've found a way to fix bug with left recursion, when the second
alternative of the head rule not tried after the first succeeded.