test: add test 21 asserting that cuts can change the grammar meaning
If some alternative that starts with a term contains a cut after that term,
then the other alternatives of the same rule (or alternatives of other rules
that can be tested on the same place) that start with the same term are not
tried. The first alternative's cut wipes backtracking data and asserts that
this rule must succeed, thus making other rules unreachable.